Refactor: Swap the naming of Viewport and ViewportBlueprint
